+++ src/hal/impl/shared/frameworks/jpeg/tjpgd.c
@@ -1,0 +1,5 @@
+#include "hal/hal_config.h"
+
+#ifdef HAL_ENABLE_JPEG
+#include <tjpgd.c>
+#endif